Revert tri-color theme picker to bi-color picker and fix theme colors for tabs
Now that ChromeRefresh2023 is active, the ThemeColorPickerHandler is using a
tri-color picker and foreground/background tabs and the tab strip were not
taking on theme colors as expected.

This change creates a BraveThemeColorPickerHandler that overrides
GetChromeColors to return bi-colors and overrides SetSeedColor to correct the
theme coloring for tabs and the tab strip.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/05baea744f4111671b9e3b9baf28f8c1727f78c5

commit 05baea744f4111671b9e3b9baf28f8c1727f78c5
Author: Alison Gale <agale@chromium.org>
Date:   Wed May 15 17:38:31 2024 +0000

    Remove references to launched kChromeRefreshSecondary2023 feature

    Also cleaned up some test-only references to kChromeRefresh2023 when
    they were adjacent.

    #chrome-refresh-2023-cleanup

    Bug: 335904039